refactor: decompose user controller and routes into modular components. Create utils.
